Add files via upload
🚀 Excited to share that I've just completed an intensive data analysis using Excel project in just one day! 📊

This project was a deep dive into various aspects of data analysis and visualization. Here’s a snapshot of what I accomplished:

- Data Cleaning: Ensured accuracy and consistency across the dataset.
- Data Analysis: Analyzed footfall and sales metrics by store location, quantity ordered, and total sales.
- Data Visualization: Created compelling charts to illustrate key insights, including average sales, average order value, and average bill per person.
- Sales Trend: Examined weekend sales, monthly sales, and weekly sales patterns.
- Product Insights: Identified the top 5 products and analyzed size distribution for better inventory management.

This project not only honed my skills in Excel but also provided valuable insights into sales performance and customer behavior. Grateful for the learning experience and excited to apply these insights in future projects!